feat: add the "createFilm" method

+++ src/main/resources/schema.sql
@@ -29,7 +29,7 @@ CREATE TABLE IF NOT EXISTS films (
     film_name VARCHAR(255) NOT NULL,
     description VARCHAR(200),
     release_date DATE CHECK release_date > '1895-12-28',
-    duration REAL CHECK duration > 0,
+    duration BIGINT CHECK duration > 0,
     genre VARCHAR(30),
     rating VARCHAR(30)
 );
